requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

discord.py[voice]==2.2.2 
elevenlabslib==0.2.5
python-dotenv==1.0.0
youtube-dl==2021.12.17
requests==2.26.0
SoundFile==0.10.3.post1
sounddevice==0.4.6